Implemented VectorOfSize
- ListOfSize is a single class that handles all the VectorOfXX
responsibilities.  It wastes a little bit of memory, but it makes
maintenance of the code much simple (for instance, we won’t have to add
map and flatmap to 16 classes).
- add Vector.push() to add a single element to the end of the vector.
Added tests for it. The intention is to add a version of push that add
VECTOR_NODE_SIZE elements to we can efficiently build a vector in
larger chunks.
- made LinkList Iterable.
- added tests for LinkList and LinkLists